pyproject.toml
[project]
name = "classification-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "datasets>=4.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "schedulefree>=1.4.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "timm>=1.0.19",
    "toml>=0.10.2",
    "torch==2.8.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.11",
]

[tool.uv.sources]
torch = [
    { index = "torch-cu128" },
]

[[tool.uv.index]]
name = "torch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
